Fix Defined Account Delete button

+++ src/main/java/com/neoris/tcl/models/ViewRollupMatchFFSSHist.java
@@ -1,6 +1,7 @@
 package com.neoris.tcl.models;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 import java.text.NumberFormat;
 import java.util.Locale;
 
@@ -25,7 +26,7 @@ public class ViewRollupMatchFFSSHist implements Serializable {
 
 	private String hfmcode;
 
-	private int balance;
+	private BigDecimal balance;
 
 	private String classification;
 
@@ -43,7 +44,7 @@ public class ViewRollupMatchFFSSHist implements Serializable {
 
 	}
 
-	public ViewRollupMatchFFSSHist(Long num, Long companyid, String hfmcode, int balance, String classification,
+	public ViewRollupMatchFFSSHist(Long num, Long companyid, String hfmcode, BigDecimal balance, String classification,
 			String naturalb, String period, String periodid, String errortext, String description) {
 		super();
 		this.num = num;
@@ -82,11 +83,11 @@ public class ViewRollupMatchFFSSHist implements Serializable {
 		this.hfmcode = hfmcode;
 	}
 
-	public int getBalance() {
+	public BigDecimal getBalance() {
 		return balance;
 	}
 
-	public void setBalance(int balance) {
+	public void setBalance(BigDecimal balance) {
 		this.balance = balance;
 	}
